moved some pictures to asssets and redirected, logo is svg, slight update to readme
