+++ src/pages/thankyou/ThankPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './thankyou.css';
-import thankpng from '../../components/ThankPic.png';
+import thankpng from '../../assets/ThankPic.png';
 import socialDis from '../../assets/social-dist.jpg';
 import reopen from '../../assets/reopen.jpg';
 import symptom from '../../assets/symptom.png';
